refactor confirm as new route
