--- packages/asl-pages/pages/role/named-person-mvp/mandatory-training/content/confirm.js
@@ -1,13 +1,0 @@
-const { merge } = require('lodash');
-const baseContent = require('../../../profile/content');
-
-module.exports = merge({}, baseContent, {
-  reviewRoleApplication: 'Add role',
-  applyingFor: 'You are assigning:',
-  onBehalfOf: 'To:',
-  rcvsNumber: 'RCVS number:',
-  explanation: 'Why is this person suitable for this role?',
-  buttons: {
-    submit: 'Submit'
-  }
-});
