--- packages/asl-pages/pages/role/named-person-mvp/mandatory-training/views/confirm.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Confirm from '../../views/confirm';
-
-export default props => <Confirm action="apply" {...props} />;
