+++ packages/asl-pages/pages/role/named-person-mvp/routes.js
@@ -1,6 +1,7 @@
 const beforeYouApply = require('./before-you-apply');
 const create = require('./apply');
 const mandatoryTraining = require('./mandatory-training');
+const confirm = require('./confirm');
 
 module.exports = {
   beforeYouApply: {
@@ -16,6 +17,11 @@ module.exports = {
   mandatoryTraining: {
     path: '/mandatory-training',
     router: mandatoryTraining,
+    breadcrumb: false
+  },
+  confirm: {
+    path: '/confirm',
+    router: confirm,
     breadcrumb: false
   }
 };
